refactor(T11447): relocate Dispatcher core to @cleocode/runtime/gateway (R3-T3 · SG-RUNTIME-UNIFICATION)
Moves the transport-agnostic CQRS dispatch core out of the CLI package so the
runtime layer owns it and ANY transport adapter (CLI/MCP/RPC/HTTP) can drive it.

- Relocated packages/cleo/src/dispatch/{dispatcher,registry,lib/meta,
  middleware/pipeline}.ts → packages/runtime/src/gateway/ (mirrors the
  @cleocode/runtime/daemon subpath, R2). Type imports repointed to
  @cleocode/contracts/gateway (R3-T2); @cleocode/core imports stay (runtime
  already depends on core — NO cycle).
- New @cleocode/runtime/gateway barrel + createGatewayHandler()/GatewayHandler
  entrypoint wrapping compose(middlewares)→DomainHandler.
- Dispatcher is injection-based (config.handlers + config.middlewares), so the
  domain handlers + transport-specific middleware (audit/budget/idempotency)
  STAY in packages/cleo and are injected — only the generic core moved.
- packages/cleo/src/dispatch/{dispatcher,registry,lib/meta,middleware/pipeline}.ts
  are now thin re-export shims → the CLI adapter, getCliDispatcher, and the ~18
  registry consumers compile unchanged (zero behavior change).
- runtime package.json ./gateway export + tsup gateway/index entry.

Verified: runtime + cleo build clean (tsc) · live dispatch smoke
(tasks.exists/find through the relocated Dispatcher) OK · cleo check arch 5/5 ·
json-stream-hygiene + contracts-fan-out + cli-boundary pass · biome clean. The
R3-T1 transport-inventory golden net + existing CLI/registry/parity tests are
the regression guard (run in CI).
