+++ packages/runtime/tsup.config.ts
@@ -4,6 +4,7 @@ export default defineConfig({
   entry: {
     index: "src/index.ts",
     "daemon/index": "src/daemon/index.ts",
+    "gateway/index": "src/gateway/index.ts",
   },
   format: ["esm"],
   target: "node20",
